+++ sdk/keyvault/azure-keyvault-keys/CHANGELOG.md
@@ -12,6 +12,8 @@
 
 ### Bugs Fixed
 
+- Fixed an replay bug in challenge authentication policy. The original request is now stored at the request level instead of the client level.
+
 ### Other Changes
 
 ## 4.12.0b2 (2026-05-29)